// src/lib/api/models/LastWatchedPatchSchemaOutputType.ts
export interface LastWatchedPatchSchemaOutputType {
    /**
     * 
     * @type {string}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    posterPath: string;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    season: number;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    episode: number;
    /**
     * 
     * @type {boolean}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    isWatched?: boolean | null;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    tmdbId: number;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    atSecond: number;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    totalSeconds: number;
    /**
     * 
     * @type {string}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    episodeName: string;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    network: number;
    /**
     * 
     * @type {string}
     * @memberof LastWatchedPatchSchemaOutputType
     */
    userId: string;
}