// src/lib/api/models/LastWatchedSchemaOutputType.ts
export interface LastWatchedSchemaOutputType {
    /**
     * 
     * @type {string}
     * @memberof LastWatchedSchemaOutputType
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof LastWatchedSchemaOutputType
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof LastWatchedSchemaOutputType
     */
    posterPath: string;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedSchemaOutputType
     */
    season: number;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedSchemaOutputType
     */
    episode: number;
    /**
     * 
     * @type {boolean}
     * @memberof LastWatchedSchemaOutputType
     */
    isWatched?: boolean | null;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedSchemaOutputType
     */
    tmdbId: number;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedSchemaOutputType
     */
    atSecond: number;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedSchemaOutputType
     */
    totalSeconds: number;
    /**
     * 
     * @type {string}
     * @memberof LastWatchedSchemaOutputType
     */
    episodeName: string;
    /**
     * 
     * @type {number}
     * @memberof LastWatchedSchemaOutputType
     */
    network: number;
    /**
     * 
     * @type {string}
     * @memberof LastWatchedSchemaOutputType
     */
    userId: string;
    /**
     * 
     * @type {ApiAuthLoginPost200Response}
     * @memberof LastWatchedSchemaOutputType
     */
    user?: ApiAuthLoginPost200Response;
}